Add files via upload
clients have individual id's and to change the id of backdoor or label flip you will need to add a new attack client with a new id for it to be fully federated. Also, there is a known but when data is split that causes the server to fail to recieve the models from the clients, this error is inconsistant at best.